requirements.txt
requests
python-gnupg
jinja2
docopts
